moved StackElement to commons project
