+++ module-stack/src/main/java/ch/hsr/adv/ui/stack/logic/StackLayouter.java
@@ -4,10 +4,10 @@ import ch.hsr.adv.commons.core.logic.domain.Module;
 import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.commons.stack.logic.ConstantsStack;
+import ch.hsr.adv.commons.stack.logic.domain.StackElement;
 import ch.hsr.adv.ui.core.logic.Layouter;
 import ch.hsr.adv.ui.core.presentation.widgets.AutoScalePane;
 import ch.hsr.adv.ui.core.presentation.widgets.LabeledNode;
-import ch.hsr.adv.ui.stack.logic.domain.StackElement;
 import com.google.inject.Singleton;
 import javafx.geometry.Insets;
 import javafx.scene.layout.Pane;
